Conflicts:
	felix/bpf-gpl/bpf.h
	felix/bpf-gpl/fib.h
	felix/bpf-gpl/globals.h
	felix/bpf/libbpf/libbpf.go
	felix/bpf/libbpf/libbpf_api.h
	felix/bpf/libbpf/libbpf_stub.go
	felix/bpf/state/map.go
	felix/bpf/tc/attach.go
	felix/dataplane/linux/bpf_ep_mgr.go
	felix/fv/bpf_attach_test.go
	felix/fv/bpf_test.go

+++ felix/bpf/ut/pol_prog_test.go
@@ -2761,7 +2761,7 @@ func (p packet) StateIn() state.State {
 		preNATDstAddr = p.preNATDstAddr
 		preNATDstPort = p.preNATDstPort
 	}
-	flags := uint8(0)
+	flags := uint64(0)
 	if p.fromHostFlag {
 		flags |= polprog.FlagSrcIsHost
 	}
